Refactors libsodium initialization
Streamlines the initialization of the libsodium library
by directly using the `sodium.ready` promise instead of
multiple checks and event listeners. This ensures that
libsodium is fully initialized before any cryptographic
operations are performed. Additionally, it includes the
libsodium library directly as a local JavaScript file
to avoid dependency on an external CDN.